backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
numpy
requests
python-dotenv
gunicorn